+++ lib/rail_feeds/national_rail.rb
@@ -2,6 +2,7 @@
 
 require_relative 'national_rail/credentials'
 require_relative 'national_rail/http_client'
+require_relative 'national_rail/knowledge_base'
 
 module RailFeeds
   module NationalRail # :nodoc: