+++ lib/rail_feeds/national_rail/knowledge_base.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative 'knowledge_base/national_service_indicator'
+
+module RailFeeds
+  module NationalRail
+    module KnowledgeBase # :nodoc:
+    end
+  end
+end