+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.0.2
 
-  * Fix issue 1: OpenURI returns a StringIO not TempFile when size is under 10KB
+  * Add National rail - Knowledge base - National service indicator via HTTP.
+  * Fix issue 1: OpenURI returns a StringIO not TempFile when size is under 10KB.
 
 ## Version 0.0.1